refactor: split wechatsync bridge tests
