+++ tests/obsidian-wechat-converter_tests_README.md
@@ -2,7 +2,7 @@
 
 ## 核心功能
 
-承载 Vitest 回归测试，覆盖渲染、同步、设置、安全、UI 和构建契约；飞书同步协调器测试按基础导入、智能覆盖和图片后处理分文件维护，AI layout service 测试按基础、渲染、缓存归一化和生成 provider 分文件维护。
+承载 Vitest 回归测试，覆盖渲染、同步、设置、安全、UI 和构建契约；飞书同步协调器测试按基础导入、智能覆盖和图片后处理分文件维护，AI layout service 测试按基础、渲染、缓存归一化和生成 provider 分文件维护，Wechatsync bridge 测试按基础服务流、队列操作、协议握手、HTTP 安全和多客户端注册表分文件维护。
 
 ## 输入
